% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{check.smooth.inputs}
\alias{check.smooth.inputs}
\title{check.smooth.inputs
Checks the inputs for smoothing and converts them into the required format for the library
Internal function}
\usage{
check.smooth.inputs(xin, nlevels, transform.type, filter, thresh.type,
  thresh, hard.soft, min.level, max.level, tol)
}
\arguments{
\item{xin}{Input vector, normally a vector to be transformed into the wavelet domain}

\item{nlevels}{Number of levels of the wavelet transform to perform}

\item{transform.type}{"DWT" or "MODWT" (time ordered)}

\item{filter}{Wavelet filter to use}

\item{thresh.type}{"manual", "univ" or "cv"}

\item{thresh}{Threshold value}

\item{hard.soft}{"hard" or "soft" thresholding}

\item{min.level}{Minimum level for thresholding}

\item{max.level}{Maximum level for thresholding}

\item{tol}{Tolerance for cross validation smoothing}
}
\description{
check.smooth.inputs
Checks the inputs for smoothing and converts them into the required format for the library
Internal function
}
