% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{check.thresh.inputs}
\alias{check.thresh.inputs}
\title{check.thresh.inputs
Checks the inputs for thresholding and converts them into the required format for the library
Internal function}
\usage{
check.thresh.inputs(xwav, hard.soft, thresh, min.level, max.level)
}
\arguments{
\item{xwav}{WST object}

\item{hard.soft}{"hard" or "soft" thresholding}

\item{thresh}{Threshold value}

\item{min.level}{Minimum level for thresholding}

\item{max.level}{Maximum level for thresholding}
}
\description{
check.thresh.inputs
Checks the inputs for thresholding and converts them into the required format for the library
Internal function
}
