% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{check.trans.inputs}
\alias{check.trans.inputs}
\title{check.trans.inputs
Checks the inputs for a transformation and converts them into the required format for the library.
Internal function}
\usage{
check.trans.inputs(xin, direction, nlevels, transform.type, filter)
}
\arguments{
\item{xin}{Input vector, normally a vector to be transformed into the wavelet domain}

\item{direction}{"FWD" or "BWD"}

\item{nlevels}{Number of levels of the wavelet transform to perform}

\item{transform.type}{"DWT" or "MODWT" (time ordered)}

\item{filter}{Wavelet filter to use}
}
\description{
check.trans.inputs
Checks the inputs for a transformation and converts them into the required format for the library.
Internal function
}
