% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{CPUReconstruct}
\alias{CPUReconstruct}
\title{CPUReconstruct}
\usage{
CPUReconstruct(xwav)
}
\arguments{
\item{xwav}{WST object}
}
\value{
Returns a vector
}
\description{
Wavelet reconstruction using the CPU.
The input argument is a transformed wavelet WST Object, containing all the necessary information to do a reconstruction.
}
\seealso{
\code{\link{CPUTransform}}, \code{\link{GPUReconstruct}}, \code{\link{CPUThreshold}}, \code{\link{CPUSmooth}}
}
