% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{CPUThreshold}
\alias{CPUThreshold}
\title{CPUThreshold}
\usage{
CPUThreshold(xwav, thresh, hard.soft, min.level, max.level)
}
\arguments{
\item{xwav}{WST object}

\item{thresh}{Threshold value}

\item{hard.soft}{"hard" or "soft"}

\item{min.level}{Min level of thresholding}

\item{max.level}{Max level of thresholding, the 'primary resolution'}
}
\description{
Threshold a transformed vector using a manual threshold, with a hard or soft threshold and specifying a range of resolution levels to threshold. Uses the CPU. For further thresholding schemes, use the Smoothing functions or wavethresh.
Using the min and max levels, we have the flexibility to implement level-dependent thresholds.
}
\seealso{
\code{\link{CPUSmooth}}, \code{\link{GPUSmooth}}, \code{\link{WSTtowavethresh}}
}
