% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{filt.convertor}
\alias{filt.convertor}
\title{filt.convertor
Converts filter identifiers between the format required for R/CUDA libraries.
Internal function}
\usage{
filt.convertor(filter, toCUDA = TRUE)
}
\arguments{
\item{filter}{Filter name}
}
\description{
filt.convertor
Converts filter identifiers between the format required for R/CUDA libraries.
Internal function
}
