% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{filtlen.calculator}
\alias{filtlen.calculator}
\title{filtlen.calculator
Returns the filter length for a given filter name.
Internal function}
\usage{
filtlen.calculator(filter)
}
\arguments{
\item{filter}{Filter name}
}
\description{
filtlen.calculator
Returns the filter length for a given filter name.
Internal function
}
