% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplotting.R
\name{getCoeffLevel}
\alias{getCoeffLevel}
\title{getCoeffLevel
Get wavelet or scaling coefficients at a given level
Internal function.}
\usage{
getCoeffLevel(xw, level, coeffType = "d")
}
\arguments{
\item{xw}{WST object}

\item{level}{Level of transform}

\item{coeffType}{Detail or scaling coefficient}
}
\description{
getCoeffLevel
Get wavelet or scaling coefficients at a given level
Internal function.
}
