% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{GPUReconstruct}
\alias{GPUReconstruct}
\title{GPUReconstruct}
\usage{
GPUReconstruct(xwav)
}
\arguments{
\item{xwav}{WST object}
}
\value{
Returns a vector
}
\description{
Wavelet reconstruction using the GPU.
The input argument is a transformed wavelet WST Object, containing all the necessary information to do a reconstruction.
}
\seealso{
\code{\link{GPUTransform}}, \code{\link{CPUReconstruct}}, \code{\link{CPUThreshold}}, \code{\link{CPUSmooth}}
}
