% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{GPUSmooth}
\alias{GPUSmooth}
\title{GPUSmooth}
\usage{
GPUSmooth(xin, nlevels, transform.type, filter, thresh.type,
  thresh = NULL, hard.soft, min.level, max.level, tol = 0.01)
}
\arguments{
\item{xin}{Vector input}

\item{nlevels}{Number of levels of transform; 0 means full transform}

\item{transform.type}{"DWT" or "MODWT"}

\item{filter}{e.g. "Haar"}

\item{thresh.type}{"manual" or "cv"}

\item{thresh}{Threshold value}

\item{hard.soft}{"hard" or "soft"}

\item{min.level}{Min level of thresholding}

\item{max.level}{Max level of thresholding, the 'primary resolution'}

\item{tol}{Tolerance for CV search}
}
\description{
Wavelet smoothing of a raw input vector. This function will perform a wavelet transform, threshold using a chosen scheme and then reconstruct. This is done on the GPU without copying the memory back and forth. Supported thresholding schemes are manual and Nason's two-fold cross validation (cv). NB we don't support universal thresholding as it is not efficient on the GPU. Options are hard or soft threshold, specifying a range of resolution levels to threshold and a tolerance parameter for the CV scheme.
}
\seealso{
\code{\link{CPUThreshold}}, \code{\link{GPUSmooth}}, \code{\link{WSTtowavethresh}}
}
