% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{label.detail.scaling}
\alias{label.detail.scaling}
\title{label.detail.scaling}
\usage{
label.detail.scaling(Xwav)
}
\arguments{
\item{Xwav}{WST object}
}
\description{
Returns a vector of "d" and "s" corresponding to detail and scaling labels for the coefficients.
Internal function.
}
