% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplotting.R
\name{makeLevelNVec}
\alias{makeLevelNVec}
\title{makeLevelNVec
Makes a list indicating length of coefficient vector for each level of transform.
Internal function}
\usage{
makeLevelNVec(XW, levelList)
}
\arguments{
\item{XW}{WST object}

\item{levelList}{List of levels of transform}
}
\description{
makeLevelNVec
Makes a list indicating length of coefficient vector for each level of transform.
Internal function
}
