% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplotting.R
\name{makeLevelTVec}
\alias{makeLevelTVec}
\title{makeLevelTVec
Makes a list of 'translate' values for each wavelet coefficient.
i.e. time value corresponding to wavelet coefficients through each level.}
\usage{
makeLevelTVec(XW, levelList)
}
\arguments{
\item{XW}{WST object}

\item{levelList}{List of levels}
}
\description{
makeLevelTVec
Makes a list of 'translate' values for each wavelet coefficient.
i.e. time value corresponding to wavelet coefficients through each level.
}
