% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplotting.R
\name{plot.WST}
\alias{plot.WST}
\title{Plotting for WST objects
Plots the wavelet transform at all levels using ggplot.}
\usage{
\method{plot}{WST}(Xwav)
}
\arguments{
\item{Xwav}{WST object}
}
\description{
Plotting for WST objects
Plots the wavelet transform at all levels using ggplot.
}
