% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{prepare.wst.arg.list}
\alias{prepare.wst.arg.list}
\title{prepare.wst.arg.list
Prepares the arguments list for reconstruction or thresholding from a WST object
Internal function}
\usage{
prepare.wst.arg.list(xwav)
}
\arguments{
\item{xwav}{WST object}
}
\description{
prepare.wst.arg.list
Prepares the arguments list for reconstruction or thresholding from a WST object
Internal function
}
