% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{print.WST}
\alias{print.WST}
\title{print.WST
Print function for the WST object. Prints a brief summary of the wavelet transform.}
\usage{
\method{print}{WST}(xwav)
}
\arguments{
\item{xwav}{WST object}
}
\description{
print.WST
Print function for the WST object. Prints a brief summary of the wavelet transform.
}
