% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{return.thresh}
\alias{return.thresh}
\title{return.thresh
Puts the results into a nice structure for returning the result of thresholding.}
\usage{
return.thresh(xwav, arg.list)
}
\arguments{
\item{arglist}{Input arguments for the transform}
}
\description{
return.thresh
Puts the results into a nice structure for returning the result of thresholding.
}
