% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{return.trans}
\alias{return.trans}
\title{return.trans
Puts the results into a nice structure for returning the result of a transform.
Internal function}
\usage{
return.trans(arglist, argsin)
}
\arguments{
\item{arglist}{Input arguments for the transform}

\item{argsin}{Input arguments for the transform}
}
\description{
return.trans
Puts the results into a nice structure for returning the result of a transform.
Internal function
}
