% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{sense.convertor}
\alias{sense.convertor}
\title{sense.convertor
Converts direction identifiers between the format required for R/CUDA libraries.
Internal function}
\usage{
sense.convertor(direction, toCUDA = TRUE)
}
\arguments{
\item{direction}{Transform direction}
}
\description{
sense.convertor
Converts direction identifiers between the format required for R/CUDA libraries.
Internal function
}
