% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplotting.R
\name{signifDown}
\alias{signifDown}
\title{signifDown
Rounds down to a specified number of signficant figures.
Internal function, used for plotting labels.}
\usage{
signifDown(x, digits)
}
\arguments{
\item{x}{Number to round}

\item{digits}{Number of significant figures}
}
\description{
signifDown
Rounds down to a specified number of signficant figures.
Internal function, used for plotting labels.
}
