% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{ttype.convertor}
\alias{ttype.convertor}
\title{ttype.convertor
Converts transform types between the format required for R/CUDA libraries.
Internal function}
\usage{
ttype.convertor(transform.type, toCUDA = TRUE)
}
\arguments{
\item{transform.type}{"DWT" or "MODWT"}
}
\description{
ttype.convertor
Converts transform types between the format required for R/CUDA libraries.
Internal function
}
