% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{WSTtoDT}
\alias{WSTtoDT}
\title{WSTtoDT}
\usage{
WSTtoDT(Xwav, scaling = TRUE, forPlotting = FALSE)
}
\arguments{
\item{Xwav}{WST object}

\item{scaling}{boolean determining whether we retain scaling coefficients (default TRUE)}

\item{forPlotting}{boolean determining whether we keep the min/max values (default FALSE)}
}
\description{
Converts a WST object to a data.table, recording time (translate) and level information for each wavelet coefficient. This takes the wavelet vector out of its 'raw' form where the coefficients are unlabelled and even interleaved for the DWT, and puts them in a more user-friendly form for performing further analysis on wavelet coefficients.
Used also for plotting.
}
