% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunctions.R
\name{WSTtoWavethresh}
\alias{WSTtoWavethresh}
\title{WSTtoWavethresh}
\usage{
WSTtoWavethresh(XW, showWarnings = TRUE)
}
\arguments{
\item{XW}{WST object}

\item{showWarnings}{Boolean option to show warnings on incompatibilities in conversions}
}
\description{
Converts a WST object to a wavethresh object to use the wavethresh library. Three things to note on design differences between wavecuda and wavethresh:
- WaveCUDA does not retain scaling coefficients not required for the transform but wavethresh does
- wavethresh does not support the C6 wavelet
- wavethresh uses a different ordering for the (time-ordered) MODWT coefficients
}
